backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
